#026 #023 #007 - moved to form components; centralizat form; validare done
